build without seccomp
If the seccomp build tag is disabled, build without it.
